Change SleepDay object to store date field as Lont type
Now SleepDay object store date in database as Long - time in millis,
(date with time of 00:00 in UTC universal time). This is set for better
database operations and UTC is for operating system to not change hours
after change of local time zone, which would be useless and confusing
in this kind of data storage and analyses.